redesigning the processor class to a factory class.
this is to align with a factory design to handle creating the different
types of stat objects and to have cleaner reading code.  By doing this I
have truncated the size of what was the Processor class (now
SkaterFactory class) and requiring only a few more lines of code for
instantiating and creating my stat objects/lists.