Refactored the process for creating the skater stat objects.
Removed SkaterStats class and replaced it with the StatProcessor class
instead.  This will handle creating the maps with team names and skater
objects with their associated stats.  This is done with three maps
covering basic, advanced, and goalie stats separately.

An IPlayerStats interface was created to connect all stat classes
together, allowing for code reuse with prep and handling of stat data.