Refactor/skater stat objects and team building (#6)
* redesigning the processor class to a factory class.

this is to align with a factory design to handle creating the different
types of stat objects and to have cleaner reading code.  By doing this I
have truncated the size of what was the Processor class (now
SkaterFactory class) and requiring only a few more lines of code for
instantiating and creating my stat objects/lists.

* Refactored the process for creating the skater stat objects.

Removed SkaterStats class and replaced it with the StatProcessor class
instead.  This will handle creating the maps with team names and skater
objects with their associated stats.  This is done with three maps
covering basic, advanced, and goalie stats separately.

An IPlayerStats interface was created to connect all stat classes
together, allowing for code reuse with prep and handling of stat data.

---------

Co-authored-by: sldri <sldri@10.0.0.46>