Major changes to project and process... (#9)
I stepped away for a couple months and coming back I saw I didn't do a
good job planning out the design of this project.  So, I'm using this to
reset and plan out the flow in which this project will work.

This is a rough draft commit.  I've got the input process for the file
consumption done just with teams to test that it works.  Next will be
handling the factory component so that we can use that consumer class to
handle the other csv files for players and goalies.

I'm also using another site that generates a ton of stats and data to
use.  I've also shied away from scraping and decided to go with a Bring
Your Own Data (BYOD) approach.

End goal will be to create the project from the ground up to have a
sound OOP design.  I want to avoid the cases where I broke DRY, had
brittle/rigid code, and create something that is far more flexible.

Co-authored-by: Shane <author@my-network-settings.com>